# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "mementos-backend",
        "description": "Documentation de l'API du projet Mementos",
        "contact": {
            "name": "Support Team",
            "email": "support@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/",
            "description": "Demo API Server"
        }
    ],
    "paths": {
...
        "/categories": {
            "get": {
                "tags": [
                    "Categories"
                ],
                "summary": "Liste toutes les catégories",
                "description": "Récupère une liste de toutes les catégories disponibles (public).",
                "operationId": "getCategories",
                "responses": {
                    "200": {
                        "description": "Opération réussie",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Categorie"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erreur interne du serveur",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Une erreur interne est survenue"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/subscribe": {
            "post": {
                "tags": [
                    "Push"
                ],
                "summary": "Abonner un utilisateur aux notifications push",
                "description": "Enregistre ou met à jour l'abonnement aux notifications push pour l'utilisateur connecté.",
                "operationId": "subscribe",
                "requestBody": {
                    "description": "Données d'abonnement push requises",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "endpoint",
                                    "keys.auth",
                                    "keys.p256dh"
                                ],
                                "properties": {
                                    "endpoint": {
                                        "type": "string",
                                        "example": "https://fcm.googleapis.com/fcm/send/eABc123YZ"
                                    },
                                    "keys": {
                                        "properties": {
                                            "auth": {
                                                "type": "string",
                                                "example": "base64-encoded-auth-key"
                                            },
                                            "p256dh": {
                                                "type": "string",
                                                "example": "base64-encoded-p256dh-key"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Abonnement réussi",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erreur de validation des données fournies",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The endpoint field is required."
                                        },
                                        "errors": {
                                            "properties": {
                                                "endpoint": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "The endpoint field is required."
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erreur interne du serveur",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Erreur interne du serveur"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Categorie": {
                "title": "Categorie",
                "description": "Schéma d'une catégorie",
                "required": [
                    "id",
                    "nom",
                    "pathImage",
                    "couleur"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "nom": {
                        "type": "string",
                        "example": "Science"
                    },
                    "pathImage": {
                        "type": "string",
                        "example": "image.jpg"
                    },
                    "couleur": {
                        "type": "string",
                        "example": "#FFFFFF"
                    },
                    "deleted_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "null"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2021-01-01T00:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2021-01-02T00:00:00Z"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}